wireless/bcmdhd: option to override suspend dtim_skip (sysfs)
When suspended, a certain amount of DTIMs are skipped in order to
save battery. The downside is, however, that e.g. push messages
are received with some delay. The override forces the wireless
adapter to pass all DTIMs.

The sysfs entry can be used to switch DTIM skip overriding on and
off. Also, it is possible to define which nth number of DTIMs is
picked up.

The sysfs path is:
/sys/module/bcmdhd/parameters/dtim_skip_override

Setting the sysfs variable to 0 disables the override and forces
the wireless driver to use the standard behavior. Setting it to 1
disables DTIM skipping and forces the driver to process all DTIMs.
Setting it to any value above 1, e.g. to 3, means that every third
DTIM request is processed.

Default value is set to 0.

Conflicts:
	drivers/net/wireless/bcmdhd/dhd_linux.c

Conflicts:
	drivers/net/wireless/bcmdhd/dhd_linux.c